# Improve RleDecoder performance