# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->
Closes #7206


# Rationale for this change

<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

Just some  small performance improvement:

```
arrow_array_reader/Int64Array/dictionary encoded, mandatory, no NULLs
                        time:   [68.938 µs 70.463 µs 72.200 µs]
                        change: [-22.970% -21.331% -19.572%] (p = 0.00 < 0.05)
                        Performance has improved.
arrow_array_reader/Int64Array/dictionary encoded, optional, no NULLs
                        time:   [77.723 µs 79.461 µs 81.282 µs]
                        change: [-22.811% -20.776% -18.535%] (p = 0.00 < 0.05)
                        Performance has improved.
```

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->



# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please call them out.
-->